agent/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
web3>=7.0.0
anthropic>=0.40.0
pandas>=2.0.0
pandas-ta>=0.3.14b1
python-dotenv>=1.0.0
aiohttp>=3.9.0
pydantic>=2.0.0
